Reduce `FormattingData` field sizes to `u16`

+++ core/src/formatter.rs
@@ -952,7 +952,7 @@ mod tests {
                     .get_formatting_data_mut(first_line_token)
                     .unwrap();
                 formatting_data.spaces_before = 0;
-                formatting_data.indentations_before = index;
+                formatting_data.indentations_before = index as u16;
             }
         }
     }
